requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.4.0

# Diffusion Models
diffusers>=0.21.0
accelerate>=0.20.0
transformers>=4.30.0

# Training & Optimization
timm>=0.9.0
einops>=0.6.1
scipy>=1.10.0
scikit-learn>=1.3.0

# Evaluation
clean-fid>=0.1.35
pytorch-fid>=0.3.0
lpips>=0.1.4

# Utilities
tqdm>=4.65.0
omegaconf>=2.3.0
wandb>=0.15.0
tensorboard>=2.13.0

# Image Processing
opencv-python>=4.7.0
imageio>=2.31.0

# Distributed Training
mpi4py>=3.1.4
